Moved UILabel extension to separate file
It didn’t like it when I tried to include it with TestableUtils, and I
don’t think it really fits there anyway.
