+++ The Mac Weekly/utils.swift
@@ -11,23 +11,6 @@ import Foundation
 import Result
 import Kingfisher
 
-
-// Source: https://stackoverflow.com/a/30711288
-extension UILabel {
-    func setHTMLFromString(htmlText: String) {
-        let modifiedFont = NSString(format:"<span style=\"font-family: '-apple-system', 'HelveticaNeue'; font-size: \(self.font!.pointSize)\">%@</span>" as NSString, htmlText) as String
-        
-        //process collection values
-        let attrStr = try! NSAttributedString(
-            data: modifiedFont.data(using: .unicode, allowLossyConversion: true)!,
-            options: [NSAttributedString.DocumentReadingOptionKey.documentType:NSAttributedString.DocumentType.html, NSAttributedString.DocumentReadingOptionKey.characterEncoding: String.Encoding.utf8.rawValue],
-            documentAttributes: nil)
-        
-        
-        self.attributedText = attrStr
-    }
-}
-
 // It doesn't seem like this function is being used. Should we just get rid of it?
 
 //func fixShadowImage(inView view: UIView) {
